[Backend][Seeder] Viết data seeder cho user, user info, wallet, vehicle
Link task: https://trello.com/c/fN0VWBMX/12-backendseeder-vi%E1%BA%BFt-data-seeder-cho-user-user-info-wallet-vehicle

+++ seeders/20220324180326-admin-info-seeder.js
@@ -9,40 +9,40 @@ module.exports = {
                     user_id: 1,
                     avatar: 'public/images/avatars/user/default-avatar.png',
                     birthday: '2001/01/01',
-                    address: '',
-                    phone_number: '',
+                    address: 'Đăk Lắk',
+                    phone_number: '0941331115',
                     gender: 1,
                 },
                 {
                     user_id: 2,
                     avatar: 'public/images/avatars/user/default-avatar.png',
                     birthday: '2001/01/01',
-                    address: '',
-                    phone_number: '',
+                    address: 'Đà Nẵng',
+                    phone_number: '0905676905',
                     gender: 1,
                 },
                 {
                     user_id: 3,
                     avatar: 'public/images/avatars/user/default-avatar.png',
                     birthday: '2001/01/01',
-                    address: '',
-                    phone_number: '',
+                    address: 'Huế',
+                    phone_number: '0394285138',
                     gender: 0,
                 },
                 {
                     user_id: 4,
                     avatar: 'public/images/avatars/user/default-avatar.png',
                     birthday: '2001/01/01',
-                    address: '',
-                    phone_number: '',
+                    address: 'Đà Nẵng',
+                    phone_number: '0702399134',
                     gender: 1,
                 },
                 {
                     user_id: 5,
                     avatar: 'public/images/avatars/user/default-avatar.png',
                     birthday: '2001/01/01',
-                    address: '',
-                    phone_number: '',
+                    address: 'Quãng Ngãi',
+                    phone_number: '0961286147',
                     gender: 1,
                 },
             ],
@@ -51,6 +51,6 @@ module.exports = {
     },
 
     async down(queryInterface, Sequelize) {
-        await queryInterface.bulkDelete('UserInfos', null, {})
+        await queryInterface.bulkDelete('UserInfos', null, { role: 2 })
     },
 }
